sources/requirements/lambda.txt
python-jose[cryptography]==3.3.0
cachecontrol[filecache]==0.12.6
